Moving WindowInfo to separate file
